-add visitor::GetDbModel

+++ Code/WorkSpeed.Data/Models/WithProductAction.cs
@@ -9,7 +9,7 @@ namespace WorkSpeed.Data.Models
     public class WithProductAction : EmployeeAction
     {
         public Product Product { get; set; }
-        public ushort ProductQuantity { get; set; }
+        public int ProductQuantity { get; set; }
 
         public float Volume => Product.Volume;
         public float Weight => Product.Weight;
